i.0.1.1.a - Console Function Added/New Version Formatting
CONSOLE
- Added a console (~ key)
- Added the ability to auto suggest commands as the user types
- Added a "memory" system, arrow keys up/down will shift through previously used commands
- Current Commands are:
   - help
   - version
   - clear
- Added color coded console texts for various things such as error, warnings, helpful, information, and normal messages
- More commands will come
- More "touch ups" will come as well

VERSION
- Messing with Version layout
- i.X.Y.Z.a
   - i = internal
   - X = major content type
   - Y = big content type
   - Z = minor content type
   - a = what version it is

Example:
i.0.1.1.a

the 0 will turn to 1 on a full launch
the current 1 represents this content type will be, right now its gonna be behind the scenes stuff or other things like this console system, when thats all done, this will change to a 2 to represent moving onto bigger things
- the second 1 represents what I am currently working on, so in this case it is the console system
- the a represents this is my first pass at the console, when I touch up more things, such as adding art, adding more commands i want etc, this will turn to b to represent that stage of progress

Ideally in the future I will have a personal road map of sorts that will be attached to patches that I want to do, for example:

 If I finish the console system,
i.0.1.1.c (the abc isnt set in stone, its just to track where I am at in terms of progress on each feature I work in)
would become:
i.0.1.2

and this i.0.1.2.x would be a new feature I am working on adding.

If/when I get around to making this personal road map, I will include it some where for others to view. This personal road map will just be a way for me to keep focus on what I need to do in the short term.